Split runner project durability tests
